extracted auth functions to their own file
